Move TreeNode next-value logic into PL/pgSQL
The purpose of this was to solve a bug when more than 10 entries at
any given level (including the root level) were inserted. The previous
next-value logic sorted lexicographically but didn't pad values when
converted back from integers, so 1000 got sorted between 100 and 200.
This led to constraint violations when more than 10 rows were
inserted.

Because the previous Python code that generated a SQL expression to
calculate the next value was a nightmare to read, and relied on a dirty
hack overriding parts of the Django ORM's internals, I took this
opportunity to rewrite it as a PL/pgSQL function; this means it can
actually be written in an imperative manner where appropriate, without
having to do multiple round-trips to the database.